House Maintenance updates
Adds a new method to HouseMaintenance to make checking if a
house is the default house easier.

Adds some more protection against the user modifying "No House"
as well as stopping the user from attempting to create a new
house with the name "No House" as allowing that would significantly
confuse the rest of the code which assumes every user has
one house with the name "No House".

It turns out "No House" is of no region as well -- so fixed an
issue when show-ing the "No House" where it would crash attempting
to dereference the nil region.